requirements.txt
fastapi[all]==0.63.0
pydantic==1.8.1
pyjwt==2.0.1
sqlalchemy==1.4.7
